requirements.txt
psycopg2==2.9.10
pytest==7.4.3
python-dotenv==1.0.1
requests==2.32.3
telepot==12.7
